Bug 528337: [ClassDiagram] EditPolicy "RemoveOrphanViewPolicy" must be disabled on Class Diagram
https://bugs.eclipse.org/bugs/show_bug.cgi?id=528337

Set RemoveOrphanViewPolicy" on deprecated and disabled it.
Apply Papyrus refactor on the custom and srcgen code

+++ plugins/uml/diagram/org.eclipse.papyrus.uml.diagram.clazz/custom-src/org/eclipse/papyrus/uml/diagram/clazz/custom/figure/ContainmentFigure.java
@@ -26,8 +26,8 @@ public class ContainmentFigure extends RoundedCompartmentFigure {
 		super.paint(graphics);
 		graphics.pushState();
 		// Draw the cross.
-		int x = (int) (bounds.width / 2);
-		int y = (int) (bounds.height / (2));
+		int x = bounds.width / 2;
+		int y = bounds.height / (2);
 		graphics.drawLine(bounds.getCenter().translate(x, 0), bounds.getCenter().translate(-x, 0));
 		graphics.drawLine(bounds.getCenter().translate(0, -y), bounds.getCenter().translate(0, y));
 		graphics.popState();
